Move data table toolbar outside the table element
